pyproject.toml
[project]
name = "ai-gateway"
version = "0.1.0"
description = "Unified AI Gateway with LiteLLM and FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "litellm",
    "pydantic",
    "jinja2",
    "httpx",
    "python-multipart",
    "redis",
    "langfuse==2.59.7",
    "python-dotenv",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.22.0",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["app"]
